Move the adapters to their own package
