+++ src/main/java/com/sen3006/coursewise/json/UserPassword.java
@@ -1,0 +1,16 @@
+package com.sen3006.coursewise.json;
+
+import com.sen3006.coursewise.models.User;
+
+public class UserPassword extends User {
+    String password;
+
+    public UserPassword(int id, String name, String surname, String email, String password) {
+        super(id, name, surname, email);
+        this.password = password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
